Clean up most circular import
The command used to check:

```
pylint --disable=all --enable=R0401 lib/ramble/ramble/*
```
