+++ lib/ramble/ramble/software_environments.py
@@ -9,7 +9,6 @@
 from collections import defaultdict
 
 import ramble.repository
-import ramble.workspace
 import ramble.keywords
 import ramble.error
 import ramble.renderer
