+++ lib/ramble/ramble/test/cmd/software_definitions.py
@@ -6,7 +6,8 @@
 # option. This file may not be copied, modified, or distributed
 # except according to those terms.
 
-from ramble.main import RambleCommand, RambleCommandError
+from ramble.main import RambleCommand
+from ramble.error import RambleCommandError
 
 software_defs = RambleCommand("software-definitions")
