+++ lib/ramble/ramble/test/end_to_end/missing_required_dry_run.py
@@ -13,7 +13,8 @@ import pytest
 import ramble.workspace
 import ramble.config
 import ramble.software_environments
-from ramble.main import RambleCommand, RambleCommandError
+from ramble.main import RambleCommand
+from ramble.error import RambleCommandError
 
 
 # everything here uses the mock_workspace_path
